// packages/data/src/generated/varka.gen.ts
export const VarkaTalents = {
  s1: {
    p1: tt([65.4598, 70.7879, 76.116, 83.7276, 89.0557, 95.145, 103.5178, 111.8905, 120.2633, 129.3972, 138.5311, 147.665, 156.799, 165.9329, 175.0668]),
    p2: tt([44.5501, 48.1762, 51.8024, 56.9826, 60.6088, 64.753, 70.4513, 76.1495, 81.8478, 88.0641, 94.2804, 100.4967, 106.7129, 112.9292, 119.1455]),
    p3: tt([23.9885, 25.941, 27.8936, 30.683, 32.6355, 34.867, 37.9353, 41.0036, 44.0719, 47.4191, 50.7664, 54.1136, 57.4608, 60.808, 64.1553]),
    p4: tt([60.239, 65.1421, 70.0453, 77.0498, 81.953, 87.5566, 95.2616, 102.9666, 110.6716, 119.077, 127.4824, 135.8879, 144.2933, 152.6988, 161.1042]),
    p5: tt([32.4364, 35.0765, 37.7167, 41.4884, 44.1285, 47.1459, 51.2947, 55.4435, 59.5924, 64.1184, 68.6444, 73.1704, 77.6964, 82.2224, 86.7484]),
    p6: tt([55.4316, 59.9434, 64.4553, 70.9008, 75.4127, 80.5691, 87.6592, 94.7493, 101.8394, 109.574, 117.3086, 125.0433, 132.7779, 140.5126, 148.2472]),
    p7: tt([29.8478, 32.2772, 34.7067, 38.1774, 40.6068, 43.3834, 47.2011, 51.0188, 54.8366, 59.0014, 63.1662, 67.331, 71.4958, 75.6606, 79.8254]),
    p8: tt([69.7498, 75.4271, 81.1044, 89.2148, 94.8921, 101.3805, 110.302, 119.2235, 128.145, 137.8775, 147.61, 157.3425, 167.0751, 176.8076, 186.5401]),
    p9: tt([37.5576, 40.6146, 43.6716, 48.0388, 51.0958, 54.5895, 59.3934, 64.1973, 69.0011, 74.2417, 79.4823, 84.7229, 89.9635, 95.2041, 100.4447]),
    p10: tt([85.6388, 92.6094, 99.58, 109.538, 116.5086, 124.475, 135.4288, 146.3826, 157.3364, 169.286, 181.2356, 193.1852, 205.1348, 217.0844, 229.034]),
    p11: tt([46.1132, 49.8666, 53.62, 58.982, 62.7354, 67.025, 72.9232, 78.8214, 84.7196, 91.154, 97.5884, 104.0228, 110.4572, 116.8916, 123.326]),
    p12: tt([50]),
    p13: tt([74.5878, 80.6589, 86.73, 95.403, 101.4741, 108.4125, 117.9528, 127.4931, 137.0334, 147.441, 157.8486, 168.2562, 178.6638, 189.0714, 199.479]),
    p14: tt([149.144, 161.2836, 173.4233, 190.7656, 202.9052, 216.7791, 235.8556, 254.9322, 274.0087, 294.8195, 315.6303, 336.4411, 357.2519, 378.0627, 398.8735]),
    p15: tt([186.2889, 201.452, 216.615, 238.2765, 253.4396, 270.7688, 294.5964, 318.4241, 342.2517, 368.2455, 394.2393, 420.2331, 446.2269, 472.2207, 498.2145]),
  },
  s2: {
    p1: tt([278.4, 299.28, 320.16, 348, 368.88, 389.76, 417.6, 445.44, 473.28, 501.12, 528.96, 556.8, 591.6, 626.4, 661.2]),
    p2: tt([12]),
    p3: tt([81.8247, 88.4848, 95.145, 104.6595, 111.3196, 118.9312, 129.3972, 139.8631, 150.3291, 161.7465, 173.1639, 184.5813, 195.9987, 207.4161, 218.8335]),
    p4: tt([55.6876, 60.2203, 64.753, 71.2283, 75.761, 80.9413, 88.0641, 95.1869, 102.3097, 110.0801, 117.8505, 125.6208, 133.3912, 141.1615, 148.9319]),
    p5: tt([29.9856, 32.4263, 34.867, 38.3537, 40.7944, 43.5837, 47.4191, 51.2545, 55.0899, 59.2739, 63.4579, 67.642, 71.826, 76.0101, 80.1941]),
    p6: tt([75.2987, 81.4277, 87.5566, 96.3123, 102.4413, 109.4458, 119.077, 128.7082, 138.3395, 148.8463, 159.3531, 169.8599, 180.3666, 190.8734, 201.3802]),
    p7: tt([40.5455, 43.8457, 47.1459, 51.8605, 55.1607, 58.9323, 64.1184, 69.3044, 74.4905, 80.148, 85.8055, 91.463, 97.1205, 102.778, 108.4355]),
    p8: tt([69.2894, 74.9293, 80.5691, 88.626, 94.2659, 100.7114, 109.574, 118.4366, 127.2992, 136.9675, 146.6358, 156.3041, 165.9724, 175.6407, 185.309]),
    p9: tt([37.3097, 40.3465, 43.3834, 47.7217, 50.7585, 54.2292, 59.0014, 63.7736, 68.5457, 73.7517, 78.9577, 84.1637, 89.3698, 94.5758, 99.7818]),
    p10: tt([87.1872, 94.2839, 101.3805, 111.5186, 118.6152, 126.7256, 137.8775, 149.0293, 160.1812, 172.3469, 184.5125, 196.6782, 208.8438, 221.0095, 233.1751]),
    p11: tt([46.947, 50.7682, 54.5895, 60.0485, 63.8697, 68.2369, 74.2417, 80.2466, 86.2514, 92.8022, 99.3529, 105.9036, 112.4544, 119.0051, 125.5559]),
    p12: tt([107.0485, 115.7618, 124.475, 136.9225, 145.6357, 155.5937, 169.286, 182.9783, 196.6705, 211.6075, 226.5445, 241.4815, 256.4185, 271.3555, 286.2925]),
    p13: tt([57.6415, 62.3333, 67.025, 73.7275, 78.4192, 83.7813, 91.154, 98.5267, 105.8995, 113.9425, 121.9855, 130.0285, 138.0715, 146.1145, 154.1575]),
    p14: tt([175.76, 188.942, 202.124, 219.7, 232.882, 246.064, 263.64, 281.216, 298.792, 316.368, 333.944, 351.52, 373.49, 395.46, 417.43]),
    p15: tt([94.64, 101.738, 108.836, 118.3, 125.398, 132.496, 141.96, 151.424, 160.888, 170.352, 179.816, 189.28, 201.11, 212.94, 224.77]),
    p16: tt([93.6, 100.62, 107.64, 117, 124.02, 131.04, 140.4, 149.76, 159.12, 168.48, 177.84, 187.2, 198.9, 210.6, 222.3]),
    p17: tt([50.4, 54.18, 57.96, 63, 66.78, 70.56, 75.6, 80.64, 85.68, 90.72, 95.76, 100.8, 107.1, 113.4, 119.7]),
    p18: tt([11]),
    p19: tt([16]),
    p20: tt([8]),
  },
  s3: {
    p1: tt([336.96, 362.232, 387.504, 421.2, 446.472, 471.744, 505.44, 539.136, 572.832, 606.528, 640.224, 673.92, 716.04, 758.16, 800.28]),
    p2: tt([181.44, 195.048, 208.656, 226.8, 240.408, 254.016, 272.16, 290.304, 308.448, 326.592, 344.736, 362.88, 385.56, 408.24, 430.92]),
    p3: tt([15]),
    p4: tt([60]),
  },
};

export const VarkaStatTable: readonly StatTableEntry[] = [
  entry("crit_dmg_base", 50, undefined, new StatTable([0, 9.6, 19.2, 19.2, 28.8, 38.4])),
  entry("crit_rate_base", 5),
  entry("recharge_base", 100),
  entry("hp_base", 981.9196, s5hp, new StatTable([841.9073, 1440.1045, 2237.701, 2835.8982, 3434.0955, 4032.2927])),
  entry("atk_base", 27.4645, s5atk, new StatTable([23.5466, 40.2771, 62.5844, 79.3149, 96.0454, 112.7759])),
  entry("def_base", 61.9241, s5hp, new StatTable([53.0955, 90.8213, 141.1223, 178.848, 216.5737, 254.2995])),
];